Rename user icon component
