+++ src/components/stats/discord/user-icon.svelte
